pyproject.toml
[project]
name = "solar_image_processing"
version = "0.1.0"
requires-python = ">=3.10,<3.13"
dependencies = [
    "numpy",
    "scipy",
    "pandas",
    "matplotlib",
    "scikit-image",
    "drms",
    "sunpy",
    "aiapy",
    "astropy",
    "joblib>=1.5.3",
    "cupy-cuda12x>=14.0.1",
    "pyyaml",
]

[project.optional-dependencies]
dev = ["pytest"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.uv]
package = true

[tool.setuptools.packages.find]
where = ["src"]

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"
